googlesheet_call/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
pandas==2.0.3
numpy==1.24.3
google-auth==2.21.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.2.0
google-api-python-client==2.153.0
httplib2==0.22.0
python-dotenv==1.0.0
requests==2.28.2
gunicorn==20.1.0
